requirements.txt
requests==2.28.1
pyyaml==6.0
schedule==1.1.0